Check that $row is not a string or array
Bypass the $row->has() test in the case that the $row is actually a string or an array. The rules checker ExistsIn method will return the row as an array, the “list” finder will return the $row as a string. Both will return a fatal error when $row->has() is called in the Crypt Behavior.